Applied filters moved to the right hand side; only one filter group expandable at a time now to reduce confusion
