+++ public/stylesheets/browser.less
@@ -2,11 +2,25 @@
   position: relative;
 }
 
+.applied-filters {
+  .list-inline {
+    margin-bottom: 0.5rem;
+  }
+
+  .badge {
+    a, a:hover, a:focus, a:visited {
+      color: #fff;
+      text-decoration: none;
+    }
+  }
+}
+
 .browser-results-container {
   li {
     margin-bottom: 20px;
   }
 }
+
 .search-result-container {
   padding: 10px 0 10px 0;
   border: 1px solid #eee;
